#3675 Notification Navigation to KiwixReaderFragment
